// app/lucy/src/app/models/Monitor.ts
/*
 * Copyright © 2019 Province of British Columbia
 * Licensed under the Apache License, Version 2.0 (the "License")
 * You may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 * **
 * http://www.apache.org/licenses/LICENSE-2.0
 * **
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 * File: Monitor.ts
 * Project: lucy
 * File Created: Tuesday, 21st January 2020 4:09:28 pm
 * Author: Williams, Andrea IIT (you@you.you)
 * -----
 * Last Modified: Tuesday, 21st January 2020 4:09:35 pm
 * Modified By: Williams, Andrea IIT (you@you.you>)
 * -----
 */


export interface EfficacyCodes {
    description: string;
    activeIndicator: boolean;
    efficacy_code_id: number;
    efficacyRating: string;
}